+++ src/Pages/Profile/Docu.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Docu = () => {
+  return (
+    <div>Docu</div>
+  )
+}
+
+export default Docu
